coursework/45_Machine_Learning_example/flask_app_classifier/requirements.txt
Pillow
Flask
Flask-Uploads 
Flask-WTF 
WTForms